+++ test/translator/function_test.exs
@@ -728,5 +728,76 @@ defmodule ElixirScript.Translator.Function.Test do
     assert_translation(ex_ast, js_code)
   end
 
+  should "translate capture operator" do
+    ex_ast = quote do
+      fun = &Kernel.is_atom/1
+    end
+
+    js_code = """
+      let fun = Kernel.is_atom;
+    """
+
+    assert_translation(ex_ast, js_code)
+
+
+    ex_ast = quote do
+      fun = &is_atom(&1)
+    end
+
+    js_code = """
+     let fun = function () {
+         return Kernel.is_atom(arguments[0]);
+     };
+    """
+
+    assert_translation(ex_ast, js_code)
+
+
+    ex_ast = quote do
+      fun = &local_function/1
+    end
+
+    js_code = """
+      let fun = local_function;
+    """
+
+    assert_translation(ex_ast, js_code)
+
+    ex_ast = quote do
+      fun = &(&1 * 2)
+    end
+
+    js_code = """
+     let fun = function () {
+         return arguments[0] * 2;
+     };
+    """
+
+    assert_translation(ex_ast, js_code)
+
+    ex_ast = quote do
+      fun = &{&1, &2}
+    end
+
+    js_code = """
+     let fun = function () {
+         return Tuple(arguments[0], arguments[1]);
+     };
+    """
+
+    assert_translation(ex_ast, js_code)
+
+    ex_ast = quote do
+      fun = &{&1, &2, &3}
+    end
+
+    js_code = """
+     let fun = function () {
+         return Tuple(arguments[0], arguments[1], arguments[2]);
+     };
+    """
+
+    assert_translation(ex_ast, js_code)
+  end
   
 end